# Conflicts:
#	lib/di/injector.config.dart
#	lib/presentation/screens/home/widgets/pop_up_start_tracking.dart

+++ lib/entities/bed_entity/bed_entity.dart
@@ -22,8 +22,8 @@ class BedEntity with _$BedEntity {
     required double resilience,
     required int time,
     required int isMint,
-    int? insurancePercent,
-    int? startTime,
-    int? endTime,
+    double? insurancePercent,
+    double? startTime,
+    double? endTime,
   }) = _BedEntity;
 }
